Fixed variable type in post_processing file

+++ src/matUtils/post_processing.hpp
@@ -10,7 +10,7 @@ void read_csv(std::unordered_map<std::string, std::string> &, const std::string)
 
 int mutation_distance(std::vector<MAT::Mutation>, std::vector<MAT::Mutation>);
 
-void compute_distance(const MAT::Tree &, const std::unordered_map<int, struct read_info*> &, const std::vector<std::string> &);
+void compute_distance(const MAT::Tree &, const std::unordered_map<size_t, struct read_info*> &, const std::vector<std::string> &);
 
 void compute_abundance(const std::unordered_map<std::string, double> &, const std::unordered_map<std::string, std::string> &);
